pyproject.toml
[project]
name = "fastapi-multi-server-mcp"
version = "0.1.0"
description = "Multi-Server MCP FastAPI Application"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "fastapi>=0.115.13",
    "mcp>=1.9.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.3",
]
